Bugfix crashing when exiting during jump
If the user exits the game session while jumping with the jump visualization plugin on, the engine will crash because the timer to collect data is not stopped. This fixes it.

Also deleted all the jump data that was saved because the data in some of them was corrupted due to this bug.